requirements.txt
numpy
gradio
opencv-python
torch
torchvision
pillow
mediapipe
rembg
matplotlib